update and rename UI: fix bug fThongTinCaNhan
- rename:
  +  fLogin to fDangNhap
  + fAccountProfile to fThongTinCaNhan
- add:
  + CheckPassword: xử lý Thread khi để mật khẩu mặc định
- update:
  + DialogDatePicker: tăng font size spinYear
  + kDatePicker: format code
