+++ src/DAO/KhachHangDAO.java
@@ -2,7 +2,6 @@ package DAO;
 
 import java.sql.*;
 import entity.KhachHang;
-import entity.NhanVien;
 
 public class KhachHangDAO {
     private static KhachHangDAO instance = new KhachHangDAO();
@@ -11,6 +10,25 @@ public class KhachHangDAO {
         if (instance == null)
             instance = new KhachHangDAO();
         return instance;
+    }
+
+    /**
+     * chưa xong
+     * @param maKH
+     * @return
+     */
+    public KhachHang getDSKhachHangByMaKH(String maKH) {
+        KhachHang data = null;
+        String query = "SELECT * FROM dbo.KhachHang kh WHERE kh.maKH == ?";
+        Object[] parameter = new Object[] { maKH };
+        ResultSet rs = DataProvider.getInstance().ExecuteQuery(query, parameter);
+        try {
+            rs.next();
+            data = new KhachHang(rs);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return data;
     }
 
     public KhachHang getKhachHangByMaKH(String maKH) {
@@ -27,13 +45,26 @@ public class KhachHangDAO {
         return data;
     }
 
-    public boolean insertKhachHang(KhachHang khachHangData) {
-        return false;
+    public boolean insertKhachHang(KhachHang khachHang) {
+        String query = "INSERT INTO dbo.KhachHang (maKH, cmnd, hoTen, gioiTinh, soDienThoai, ngaySinh) VALUES ( ? , ? , ? , ? , ? , ? )";
+        Object[] parameter = new Object[] { khachHang.getMaKH(), khachHang.getCmnd(), khachHang.getHoTen(),
+                khachHang.getGioiTinh(), khachHang.getSoDienThoai(), khachHang.getNgaySinh() };
+        int result = DataProvider.getInstance().ExecuteNonQuery(query, parameter);
+        return result > 0;
     }
-    
+
     public String getMaKHCuoiCung() {
-		String maKH = "";
-		
-		return maKH;
-	}
+        String query = "SELECT TOP 1 * FROM dbo.KhachHang ORDER BY maKH DESC";
+        Object[] parameter = new Object[] {};
+        String maKH = (String) DataProvider.getInstance().ExecuteScalar(query, parameter);
+        return maKH;
+    }
+
+    public boolean updateKhachHang(KhachHang khachHang) {
+        String query = "Update dbo.KhachHang set cmnd = ? , hoTen = ? , gioiTinh = ? , soDienThoai = ? , ngaySinh = ? , Where maKH = ?";
+        Object[] parameter = new Object[] { khachHang.getCmnd(), khachHang.getHoTen(), khachHang.getGioiTinh(),
+                khachHang.getSoDienThoai(), khachHang.getNgaySinh(), khachHang.getMaKH() };
+        int result = DataProvider.getInstance().ExecuteNonQuery(query, parameter);
+        return result > 0;
+    }
 }
